RoomFeature roomType was connected with RoomClassification roomType

+++ src/main/java/com/example/yurt2/Entities/RoomClassification.java
@@ -40,7 +40,7 @@ public class RoomClassification {
         this.roomType = roomType;
     }
 
-    public int getDormitoryId() {
+    public long getDormitoryId() {
         return dormitoryId;
     }
